handle upgrade call
